// src/common/models/Property.tsx
export type Property = {
  BedroomsTotal: number;
  OnMarketDate: string;
  Flooring: string[];
  PhotosCount: number;
  FireplacesTotal: number;
  Cooling: string[];
  GarageYN: boolean;
  ExteriorFeatures: string[];
  Longitude: number;
  PublicRemarks: string;
  RoomBathroomFeatures: string[];
  RoomDiningRoomFeatures: string[];
  Appliances: string[];
  TaxAnnualAmount: number;
  ListOfficeName: string;
  Basement: string[];
  Latitude: number;
  PropertyType: string;
  BathroomsTotalDecimal: number;
  ListPrice: number;
  MlsStatus: string;
  RoomMasterBedroomFeatures: string[];
  Gas: string[];
  FireplaceYN: boolean;
  AttachedGarageYN: boolean;
  StandardStatus: string;
  Roof: string[];
  Stories: number;
  Levels: string[];
  InteriorFeatures: string[];
  BathroomsFull: number;
  LotSizeAcres: number;
  SubdivisionName: string;
  Fencing: string[];
  WaterSource: string[];
  LotSizeDimensions: string;
  StoriesTotal: number;
  YearBuilt: number;
  RoomKitchenFeatures: string[];
  OtherStructures: string[];
  LivingArea: number;
  LivingAreaUnits: string;
  ParkingFeatures: string[];
  BathroomsHalf: number;
  LotSizeArea: number;
  Sewer: string[];
  Heating: string[];
  PatioAndPorchFeatures: string[];
  Utilities: string[];
  FireplaceFeatures: string[];
  WindowFeatures: string[];
  CommunityFeatures: string[];
  RoomsTotal: number;
  GarageSpaces: number;
  Media: PropertyMedia[];
  City: string;
  TaxYear: number;
  Ownership: string;
  LotFeatures: string[];
  PostalCode: string;
  LotSizeSquareFeet: number;
  SrchitecturalStyle: string[];
  UnparsedAddress: string;
  BuildingAreaTotal: number;
  ListingId: string;
  PropertySubType: string;
  DaysOnMarket: number;
  AssociationFee: number;
};

export type PropertyMedia = {
  Order: number;
  MediaURL: string;
  ClassName: string;
  ResourceName: string;
  MediaCategory: string;
  MimeType: string;
  ShortDescription: string;
};